# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ProfileBot API
  description: API per gestione embedding e servizi di salute applicativa.
  contact:
    name: ProfileBot Team
    email: team@profilebot.local
  version: 0.1.0
servers:
  - url: /
    description: Default
paths:
...
  /api/v1/availability/load:
    post:
      tags:
        - availability
      summary: Carica disponibilità da CSV
      description: Load availability data into Redis cache from CSV path or body.
      operationId: load_availability_api_v1_availability_load_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityLoadRequest"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityLoadResponse"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "503":
          description: Redis unavailable
  /api/v1/availability/{res_id}:
    get:
      tags:
        - availability
      summary: Ottieni disponibilità per res_id
      description: Return availability record for a given res_id from cache.
      operationId: get_availability_api_v1_availability__res_id__get
      parameters:
        - name: res_id
          in: path
          required: true
          schema:
            type: integer
            title: Res Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
        "404":
          description: Not Found
        "503":
          description: Redis unavailable
  /api/v1/availability/stats:
    get:
      tags:
        - availability
      summary: Statistiche disponibilità cache
      description: Return availability cache statistics.
      operationId: get_availability_stats_api_v1_availability_stats_get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityStatsResponse"
        "503":
          description: Redis unavailable
components:
  schemas:
...
    AvailabilityLoadRequest:
      properties:
        csv_path:
          anyOf:
            - type: string
            - type: "null"
          title: Csv Path
        csv_content:
          anyOf:
            - type: string
            - type: "null"
          title: Csv Content
      type: object
      title: AvailabilityLoadRequest
    AvailabilityLoadResponse:
      properties:
        total_rows:
          type: integer
          title: Total Rows
        loaded:
          type: integer
          title: Loaded
        skipped:
          type: integer
          title: Skipped
      type: object
      required:
        - total_rows
        - loaded
        - skipped
      title: AvailabilityLoadResponse
    AvailabilityResponse:
      properties:
        res_id:
          type: integer
          title: Res Id
        status:
          type: string
          enum:
            - free
            - partial
            - busy
            - unavailable
          title: Status
        allocation_pct:
          type: integer
          minimum: 0.0
          maximum: 100.0
          title: Allocation Pct
        current_project:
          anyOf:
            - type: string
            - type: "null"
          title: Current Project
        available_from:
          anyOf:
            - type: string
            - type: "null"
          title: Available From
        updated_at:
          type: string
          title: Updated At
      type: object
      required:
        - res_id
        - status
        - allocation_pct
        - updated_at
      title: AvailabilityResponse
    AvailabilityStatsResponse:
      properties:
        total:
          type: integer
          title: Total
        by_status:
          type: object
          additionalProperties:
            type: integer
          title: By Status
        last_updated_at:
          anyOf:
            - type: string
            - type: "null"
          title: Last Updated At
      type: object
      required:
        - total
        - by_status
      title: AvailabilityStatsResponse
...